+++ modules/page-help.php
@@ -6,55 +6,42 @@ register_module([
 	"description" => "Adds the help action. You really want this one.",
 	"id" => "page-help",
 	"code" => function() {
+		global $settings;
+		
 		add_action("help", function() {
-			global $settings, $version;
+			global $settings, $version, $help_sections;
 			
 			$title = "Help - $settings->sitename";
+			
+			// Sort the help sections by key
+			ksort($help_sections, SORT_NATURAL);
+			
+			$content .= "	<h1>$settings->sitename Help</h1>
+		<p>Welcome to $settings->sitename!</p>
+		<p>$settings->sitename is powered by Pepperminty wiki, a complete wiki in a box you can drop into your server.</p>";
+			
+			// todo Insert a tabel of contents here?
+			
+			foreach($help_sections as $index => $section)
+			{
+				// Todo add a button that you can click to get a permanent link
+				// to this section.
+				$content .= "<h2 id='$index'>" . $section["title"];
+				$content .= $section["content"] . "\n";
+			}
+			
 			$content = "	<h1>$settings->sitename Help</h1>
-	<p>Welcome to $settings->sitename!</p>
-	<p>$settings->sitename is powered by Pepperminty wiki, a complete wiki in a box you can drop into your server.</p>
-	<h2>Navigating</h2>
-	<p>All the navigation links can be found in the top right corner, along with a box in which you can type a page name and hit enter to be taken to that page (if your site administrator has enabled it).</p>
-	<p>In order to edit pages on $settings->sitename, you probably need to be logged in. If you do not already have an account you will need to ask $settings->sitename's administrator for an account since there is not registration form. Note that the $settings->sitename's administrator may have changed these settings to allow anonymous edits.</p>
-	<h2>Editing</h2>
-	<p>$settings->sitename's editor uses a modified version of slimdown, a flavour of markdown that is implementated using regular expressions. See the credits page for more information and links to the original source for this. A quick reference can be found below:</p>
-	<table>
-		<tr><th>Type This</th><th>To get this</th>
-		<tr><td><code>_italics_</code></td><td><em>italics</em></td></tr>
-		<tr><td><code>*bold*</code></td><td><strong>bold</strong></td></tr>
-		<tr><td><code>~~Strikethrough~~</code></td><td><del>Strikethough</del></td></tr>
-		<tr><td><code>`code`</code></td><td><code>code</code></td></tr>
-		<tr><td><code># Heading</code></td><td><h2>Heading</h2></td></tr>
-		<tr><td><code>## Sub Heading</code></td><td><h3>Sub Heading</h3></td></tr>
-		<tr><td><code>[[Internal Link]]</code></td><td><a href='index.php?page=Internal Link'>Internal Link</a></td></tr>
-		<tr><td><code>[[Display Text|Internal Link]]</code></td><td><a href='index.php?page=Internal Link'>Display Text</a></td></tr>
-		<tr><td><code>[Display text](//google.com/)</code></td><td><a href='//google.com/'>Display Text</a></td></tr>
-		<tr><td><code>&gt; Blockquote<br />&gt; Some text</code></td><td><blockquote> Blockquote<br />Some text</td></tr>
-		<tr><td><code> - Apples<br /> * Oranges</code></td><td><ul><li>Apples</li><li>Oranges</li></ul></td></tr>
-		<tr><td><code>1. This is<br />2. an ordered list</code></td><td><ol><li>This is</li><li>an ordered list</li></ol></td></tr>
-		<tr><td><code>
-	---
-	</code></td><td><hr /></td></tr>
-		<!--<tr><tds><code> - One
- - Two
- - Three</code></td><td><ul><li>One</li><li>Two</li><li>Three</li></ul></td></tr>-->
-		<tr><td><code>![Alt text](//starbeamrainbowlabs.com/favicon-small.png)</code></td><td><img src='//starbeamrainbowlabs.com/favicon-small.png' alt='Alt text' /></td></code>
-	</table>
-	
-	<p>In addition, the following extra syntax is supported for images:</p>
-	
-	<pre><code>Size the image to at most 250 pixels wide:
-![Alt text](//starbeamrainbowlabs.com/favicon-small.png 250px)
-
-Size the image to at most 120px wide and have it float at the right ahnd size of the page:
-![Alt text](//starbeamrainbowlabs.com/favicon-small.png 120px right)</code></pre>
-	
 	<h2>Administrator Actions</h2>
 	<p>By default, the <code>delete</code> and <code>move</code> actions are shown on the nav bar. These can be used by administrators to delete or move pages.</p>
 	<p>The other thing admininistrators can do is update the wiki (provided they know the site's secret). This page can be found here: <a href='?action=update'>Update $settings->sitename</a>.</p>
 	<p>$settings->sitename is currently running on Pepperminty Wiki <code>$version</code></p>";
 			exit(page_renderer::render_main($title, $content));
 		});
+		
+		// Register a help section on general navigation
+		add_help_section("5-navigation", "Navigation", "<h2>Navigating</h2>
+		<p>All the navigation links can be found on the top bar, along with a search box (if your site administrator has enabled it). There is also a &quot;More...&quot; menu in the top right that contains some additional links that you may fine useful.</p>
+		<p>This page, along with the credits page, can be found on the bar at the bottom of every page.</p>");
 	}
 ]);