+++ modules/parser-default.php
@@ -1,14 +1,52 @@
 <?php
 register_module([
 	"name" => "Default Parser",
-	"version" => "0.8",
+	"version" => "0.9",
 	"author" => "Johnny Broadway & Starbeamrainbowlabs",
 	"description" => "The default parser for Pepperminty Wiki. Based on Johnny Broadway's Slimdown (with more than a few modifications). This parser's features are documented in the help page.",
 	"id" => "parser-default",
 	"code" => function() {
+		global $settings;
+		
 		add_parser("default", function($markdown) {
 			return Slimdown::render($markdown);
 		});
+		
+		// Register the help section
+		if($settings->parser != "default")
+			return; // Don't register the help section if we aren't the currently set parser.
+		add_help_section("20-parser-default", "Editor Syntax", "<h2>Editing</h2>
+		<p>$settings->sitename's editor uses a modified version of slimdown, a flavour of markdown that is implementated using regular expressions. See the credits page for more information and links to the original source for this. A quick reference can be found below:</p>
+		<table>
+			<tr><th>Type This</th><th>To get this</th>
+			<tr><td><code>_italics_</code></td><td><em>italics</em></td></tr>
+			<tr><td><code>*bold*</code></td><td><strong>bold</strong></td></tr>
+			<tr><td><code>~~Strikethrough~~</code></td><td><del>Strikethough</del></td></tr>
+			<tr><td><code>`code`</code></td><td><code>code</code></td></tr>
+			<tr><td><code># Heading</code></td><td><h2>Heading</h2></td></tr>
+			<tr><td><code>## Sub Heading</code></td><td><h3>Sub Heading</h3></td></tr>
+			<tr><td><code>[[Internal Link]]</code></td><td><a href='index.php?page=Internal Link'>Internal Link</a></td></tr>
+			<tr><td><code>[[Display Text|Internal Link]]</code></td><td><a href='index.php?page=Internal Link'>Display Text</a></td></tr>
+			<tr><td><code>[Display text](//google.com/)</code></td><td><a href='//google.com/'>Display Text</a></td></tr>
+			<tr><td><code>&gt; Blockquote<br />&gt; Some text</code></td><td><blockquote> Blockquote<br />Some text</td></tr>
+			<tr><td><code> - Apples<br /> * Oranges</code></td><td><ul><li>Apples</li><li>Oranges</li></ul></td></tr>
+			<tr><td><code>1. This is<br />2. an ordered list</code></td><td><ol><li>This is</li><li>an ordered list</li></ol></td></tr>
+			<tr><td><code>
+		---
+		</code></td><td><hr /></td></tr>
+			<!--<tr><tds><code> - One
+	 - Two
+	 - Three</code></td><td><ul><li>One</li><li>Two</li><li>Three</li></ul></td></tr>-->
+			<tr><td><code>![Alt text](//starbeamrainbowlabs.com/favicon-small.png)</code></td><td><img src='//starbeamrainbowlabs.com/favicon-small.png' alt='Alt text' /></td></code>
+		</table>
+		
+		<p>In addition, the following extra syntax is supported for images:</p>
+		
+		<pre><code>Size the image to at most 250 pixels wide:
+	![Alt text](//starbeamrainbowlabs.com/favicon-small.png 250px)
+	
+	Size the image to at most 120px wide and have it float at the right ahnd size of the page:
+	![Alt text](//starbeamrainbowlabs.com/favicon-small.png 120px right)</code></pre>");
 	}
 ]);