+++ Changelog.md
@@ -9,4 +9,4 @@
  - Improved apparence of the tag list page.
  - Added the all tags page to the "More..." menu by default.
  - Added a link back to the list of tags on the list of pages with a particular tag.
- - 
+ - Upgrade help page. Modules can now register their own sections on a wiki's help page.